+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 Hi! This is where we document all notable changes, including bug fixes, enhancements, and dependency updates.
 Dates should be in`YYYY-MM-DD` format and versions are in [semantic versioning](http://semver.org/) format.
 
+## 0.1.10 2025-03-09
+
+### Maintenance
+
+- Updated minor versions of Playright, Svelte, SvelteKit, Vite, Vitest.
+
 ## 0.1.9 2025-03-02
 
 ### Fixed